pyproject.toml
[project]
name = "iris"
version = "0.1.0"
description = "IRIS - Video understanding ML system"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.5.1",
    "torch>=2.5.1",
    "jupyter>=1.1.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "ruff>=0.13.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "websockets>=15.0.1",
    "fastapi>=0.121.1",
    "uvicorn>=0.38.0",
    "opencv-python-headless>=4.11.0.86",
    "threaded-videocapture>=1.0.1",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.53.0",
    "scikit-learn>=1.8.0",
]

[dependency-groups]
dev = [
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
]

# Server-only dependencies (V100 + Mac dev)
server = [
    # Core ML stack - tested versions for Qwen2.5-VL
    "torch>=2.3.0,<2.6.0",
    "torchvision>=0.20.1,<0.21.0",
    "torchaudio>=2.5.1,<2.6.0",
    "transformers>=4.49.0,<4.54.0",                   # Qwen2.5-VL support, avoid 4.54+ regression
    "peft>=0.14.0",                                   # VLM support with bitsandbytes
    "trl<=0.26.1",                                    # Native VLM support in SFTTrainer
    "bitsandbytes>=0.46.0 ; sys_platform == 'linux'", # 4-bit quantization
    "accelerate>=0.21.0",                             # Distributed training
    "qwen-vl-utils>=0.0.8",                           # Qwen utilities
    "datasets>=2.19.0",                               # Data loading

    "num2words>=0.5.14",
    # Platform-specific
    "nvidia-ml-py>=11.5.0 ; platform_system == 'Linux'",
    "torchcodec>=0.7.0 ; sys_platform == 'darwin'",
    "flash-attn>=2.7.3; sys_platform == 'linux' and platform_machine == 'x86_64'",
    "scikit-learn>=1.8.0",
    # Logging
    "tensorboard>=2.14.0",
    "wandb>=0.23.1",
]

[project.scripts]
iris-train = "iris.cli.finetune.train:main"
# iris-client-dev = "iris.client.web.run_dev:main"
iris-client = "iris.cli.client:main_cli"
iris-server = "iris.cli.server:run"
iris-benchmark = "iris.cli.benchmark:run"
iris-vis = "iris.dataset.visualize_sampling:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/iris"]

# ============================================
# UV PYTORCH CONFIGURATION
# ============================================

# Define PyTorch CUDA 12.1 index
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

# Define Flash Attention pre-built wheels index
# [[tool.uv.index]]
# name = "flash-attn-wheels"
# url = "https://github.com/Dao-AILab/flash-attention/releases/download/v2.7.3/"
# explicit = true

# Point PyTorch packages to CUDA 12.1 index on Linux
# Point flash-attn to pre-built wheels to avoid building from source
[tool.uv.sources]
torch = [{ index = "pytorch-cu121", marker = "sys_platform == 'linux'" }]
torchvision = [{ index = "pytorch-cu121", marker = "sys_platform == 'linux'" }]
torchaudio = [{ index = "pytorch-cu121", marker = "sys_platform == 'linux'" }]
# flash-attn = [
#     { index = "flash-attn-wheels", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" }
# ]

[tool.uv.extra-build-dependencies]
flash-attn = [
    { requirement = "torch", match-runtime = true },
    "ninja",
    "packaging",
    "setuptools",
    "wheel",
]
